requirements.txt
django==5.0.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
redis==5.0.7
django-redis==5.4.0
requests==2.32.3
gunicorn==22.0.0
uvicorn[standard]==0.30.3
pytz==2024.1

pytest>=8.3.1
pytest-django>=4.8.0
factory-boy>=3.3.0

ipython>=8.26.0
pre_commit>=3.7.1
pip-check-updates>=0.27.0

isort>=5.13.2
bandit>=1.7.9
black>=24.4.2
prospector>=1.10.3
setuptools>=71.1.0
pyupgrade>=3.16.0
